remove `{{action}}` helper
