+++ test-app/lib/ember-blog/addon/routes/post.js
@@ -1,7 +1,6 @@
 /* We intentionally test replaceWithExternal method which is not available in controllers */
 /* eslint-disable ember/no-controller-access-in-routes */
 import Route from '@ember/routing/route';
-import { action } from '@ember/object';
 
 export default class extends Route {
   model(params) {
@@ -10,23 +9,5 @@ export default class extends Route {
       id: params.id,
       title: `Post ${params.id}`,
     };
-  }
-
-  @action goToChineseVersion() {
-    this.transitionTo({ queryParams: { lang: 'Chinese' } });
-  }
-
-  @action transitionToHome() {
-    this.transitionToExternal('home').then(() => {
-      var postController = this.controllerFor(this.routeName);
-      postController.set('transitionedToExternal', true);
-    });
-  }
-
-  @action replaceWithHome() {
-    this.replaceWithExternal('home').then(() => {
-      var postController = this.controllerFor(this.routeName);
-      postController.set('replacedWithExternal', true);
-    });
   }
 }
